undefined conversion fixed

+++ src/server/party.c
@@ -82,7 +82,7 @@ bool WriteAccount(struct account *r_acc, bool new) {
 			if (!strcmp(c_acc.name, r_acc->name)) found = 1;
 		}
 		if (found) {
-			fseek(fp, -sizeof(struct account), SEEK_CUR);
+			fseek(fp, -((signed int)sizeof(struct account)), SEEK_CUR);
 			if (fwrite(r_acc, sizeof(struct account), 1, fp) < 1) {
 				s_printf("Writing to account file failed: %s\n", feof(fp) ? "EOF" : strerror(ferror(fp)));
 			}
@@ -565,7 +565,7 @@ bool GetAccount(struct account *c_acc, cptr name, char *pass, bool leavepass) {
 				/* Update the timestamp if the password is successfully verified - mikaelh */
 				if (val == 0) {
 					c_acc->acc_laston_real = c_acc->acc_laston = time(NULL);
-					fseek(fp, -((signed long)sizeof(struct account)), SEEK_CUR);
+					fseek(fp, -((signed int)sizeof(struct account)), SEEK_CUR);
 					if (fwrite(c_acc, sizeof(struct account), 1, fp) < 1) {
 						s_printf("Writing to account file failed: %s\n", feof(fp) ? "EOF" : strerror(ferror(fp)));
 					}
@@ -674,7 +674,7 @@ bool GetcaseAccount(struct account *c_acc, cptr name, char *correct_name, bool l
 				/* Update the timestamp if the password is successfully verified - mikaelh */
 				if (val == 0) {
 					c_acc->acc_laston_real = c_acc->acc_laston = time(NULL);
-					fseek(fp, -sizeof(struct account), SEEK_CUR);
+					fseek(fp, -((signed int)sizeof(struct account)), SEEK_CUR);
 					if (fwrite(c_acc, sizeof(struct account), 1, fp) < 1) {
 						s_printf("Writing to account file failed: %s\n", feof(fp) ? "EOF" : strerror(ferror(fp)));
 					}
@@ -5094,7 +5094,7 @@ void scan_accounts() {
 
 		//if (modified) WriteAccount(&acc, FALSE);
 		if (modified) {
-			fseek(fp, -((long)(sizeof(struct account))), SEEK_CUR);
+			fseek(fp, -((signed int)(sizeof(struct account))), SEEK_CUR);
 			if (fwrite(&acc, sizeof(struct account), 1, fp) < 1) {
 				s_printf("Writing to account file failed: %s\n", feof(fp) ? "EOF" : strerror(ferror(fp)));
 			}
